Remove retrieval_coordinator module
