+++ vaannotate/vaannotate_ai_backend/services/context_builder.py
@@ -398,7 +398,7 @@ class ContextBuilder:
             oversample=pool_oversample,
         )
 
-        diagnostics["retrieval_coordinator"] = pool_diag
+        diagnostics["retrieval"] = pool_diag
         diagnostics["keyword_search"] = {
             "keywords": uniq_keywords,
             "hit_count": pool_diag.get("keyword_hits", 0),
